Update elbv2.py
Check that the attribute has been assigned and if it has that it is populated.

+++ src/acktest/bootstrapping/elbv2.py
@@ -82,7 +82,7 @@ class NetworkLoadBalancer(Bootstrappable):
   def cleanup(self):
     """Deletes a Network Load Balancer.
     """
-    if self.arn:
+    if hasattr(self, "arn") and self.arn:
       self.elbv2_client.delete_load_balancer(
         LoadBalancerArn=self.arn
       )
